+++ src/main/java/in/co/examsadda/jpa/QuestionRepository.java
@@ -1,6 +1,5 @@
 package in.co.examsadda.jpa;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -9,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import in.co.examsadda.entity.Question;
 
 @Repository
-public interface QuestionRepository extends JpaRepository<Question, BigInteger> {
+public interface QuestionRepository extends JpaRepository<Question, Long> {
 	
 	public List<Question> findAllBySectionSectionId(int sectionId);
 }
